requirements.txt
allennlp==1.2.2
allennlp-models==1.2.2
attrs==20.2.0
blis==0.4.1
boto3==1.15.17
botocore==1.18.17
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.1
conllu==4.2.1
cupy==8.1.0
cymem==2.0.3
Cython==0.29.21
dataclasses==0.6
docker-pycreds==0.4.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0a0/en_core_web_lg-3.0.0a0.tar.gz
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0a0/en_core_web_md-3.0.0a0.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0a0/en_core_web_sm-3.0.0a0.tar.gz
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.0.0a0/en_core_web_trf-3.0.0a0.tar.gz
fastrlock==0.5
filelock==3.0.12
Flask==1.1.2
ftfy==5.8
future==0.18.2
gitdb==4.0.5
GitPython==3.1.9
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
h5py==3.1.0
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonnet==0.17.0
jsonpickle==1.4.2
lemminflect==0.2.1
MarkupSafe==1.1.1
msgpack==1.0.0
murmurhash==1.0.2
ndjson==0.3.1
nltk==3.5
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
overrides==3.1.0
packaging==20.4
pathtools==0.1.2
pathy==0.3.1
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
promise==2.3
protobuf==3.13.0
psutil==5.7.2
py==1.9.0
py-rouge==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
pyparsing==3.0.0a2
pytest==6.1.1
python-dateutil==2.8.1
pytokenizations==0.6.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
sentence-transformers==0.3.9
sentencepiece==0.1.91
sentry-sdk==0.19.0
shortuuid==1.0.1
six==1.15.0
smart-open==2.2.1
smmap==3.0.4
spacy-lookups-data==0.3.2
spacy-nightly==3.0.0rc1
spacy-transformers==1.0.0a23
srsly==2.3.2
subprocess32==3.5.4
tensorboardX==2.1
thinc==8.0.0rc0
threadpoolctl==2.1.0
tokenizers==0.9.3
toml==0.10.1
torch==1.7.0
torchcontrib==0.0.2
tqdm==4.50.2
transformers==3.5.1
typer==0.3.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
wandb==0.10.7
wasabi==0.8.0
watchdog==0.10.3
wcwidth==0.2.5
Werkzeug==1.0.1
word2number==1.1
